Adding jenkins file
