+++ docker-compose.yml
@@ -1,10 +1,7 @@
-version: '2.1'
+version: '3'
 services:
   localstack:
     image: localstack/localstack
-    ports:
-      - "4567-4599:4567-4599"
-      - "${PORT_WEB_UI-8080}:${PORT_WEB_UI-8080}"
     environment:
       - SERVICES=${SERVICES- }
       - DEBUG=${DEBUG- }
@@ -13,3 +10,15 @@ services:
       - LAMBDA_EXECUTOR=${LAMBDA_EXECUTOR- }
       - KINESIS_ERROR_PROBABILITY=${KINESIS_ERROR_PROBABILITY- }
       - DOCKER_HOST=unix:///var/run/docker.sock
+  go:
+    build:
+      context: .
+      dockerfile: Dockerfile.go+terraform
+    command: >
+        bash -c "cd /code/test
+        && pwd
+        && go test"
+    volumes:
+      - .:/code
+    depends_on:
+      - localstack
