+++ examples/iam-account/main.tf
@@ -8,25 +8,25 @@ provider "aws" {
   skip_requesting_account_id  = true
 
   endpoints {
-    apigateway     = "http://localhost:4567"
-    cloudformation = "http://localhost:4581"
-    cloudwatch     = "http://localhost:4582"
-    dynamodb       = "http://localhost:4569"
-    es             = "http://localhost:4578"
-    firehose       = "http://localhost:4573"
-    iam            = "http://localhost:4593"
-    kinesis        = "http://localhost:4568"
-    lambda         = "http://localhost:4574"
-    route53        = "http://localhost:4580"
-    redshift       = "http://localhost:4577"
-    s3             = "http://localhost:4572"
-    secretsmanager = "http://localhost:4584"
-    ses            = "http://localhost:4579"
-    sns            = "http://localhost:4575"
-    sqs            = "http://localhost:4576"
-    ssm            = "http://localhost:4583"
-    stepfunctions  = "http://localhost:4585"
-    sts            = "http://localhost:4592"
+    apigateway     = "http://localstack:4567"
+    cloudformation = "http://localstack:4581"
+    cloudwatch     = "http://localstack:4582"
+    dynamodb       = "http://localstack:4569"
+    es             = "http://localstack:4578"
+    firehose       = "http://localstack:4573"
+    iam            = "http://localstack:4593"
+    kinesis        = "http://localstack:4568"
+    lambda         = "http://localstack:4574"
+    route53        = "http://localstack:4580"
+    redshift       = "http://localstack:4577"
+    s3             = "http://localstack:4572"
+    secretsmanager = "http://localstack:4584"
+    ses            = "http://localstack:4579"
+    sns            = "http://localstack:4575"
+    sqs            = "http://localstack:4576"
+    ssm            = "http://localstack:4583"
+    stepfunctions  = "http://localstack:4585"
+    sts            = "http://localstack:4592"
   }
 }
